RSOC_OS/requirements.txt
PySide6
pywin32
pyautogui
pygetwindow
pywinauto
requests
keyboard
PyMuPDF
pandas
openpyxl
python-docx